[refactor] order 코드 정리 #75
